RavenDB-19197 : cleanup + move GenerateShardRestoreSettings to Sharding.Backup
