+++ src/Raven.Server/Documents/TransactionOperationsMerger.cs
@@ -18,7 +18,6 @@ using Sparrow.Json;
 using Sparrow.Json.Sync;
 using Sparrow.Logging;
 using Sparrow.LowMemory;
-using Sparrow.Server.Json.Sync;
 using Sparrow.Server.Meters;
 using Sparrow.Server.Utils;
 using Sparrow.Utils;
