+++ src/Raven.Server/ServerWide/Context/DocumentsOperationContext.cs
@@ -2,8 +2,6 @@
 using System.Collections.Generic;
 using Raven.Server.Documents;
 using Raven.Server.Utils;
-using Sparrow.Collections;
-using Sparrow.Json;
 using Sparrow.Threading;
 using Voron;
